db/mqtt-ingestor/requirements.txt
paho-mqtt==2.1.0
psycopg[binary]==3.2.7
